Update ~^/static/ rewrite rules from last to break.
Thus the re-write processing result will be remained and stop just
inside the ~^/static/ location block. Then the request will be
redirected to the proxy pass upstream.

If the rule is last, then the request will be passed to other processors
and finally be matched later there, so the backend won't receive the request
as expected